chore:  add new columns

+++ HackerNewsScraper/HackerNewApi/HackerNewsItem.cs
@@ -6,11 +6,11 @@ namespace HackerNewsScraper.HackerNewApi
 {
     public class HackerNewsItem
     {
-        public long id { get; set; }
+        public int id { get; set; }
         public string text { get; set; }
         public string by { get; set; }
-        public long? parent { get; set; }
-        public List<long> kids { get; set; }
+        public int? parent { get; set; }
+        public List<int> kids { get; set; }
         public string type { get; set; }
         public long time { get; set; }
     }
