+++ HackerNewsScraper/HackerNewApi/HackerNewsStory.cs
@@ -6,10 +6,10 @@ namespace HackerNewsScraper.HackerNewApi
 {
     public class HackerNewsStory
     {
-        public long id { get; set; }
+        public int id { get; set; }
         public string title { get; set; }
         public string by { get; set; }
-        public List<long> kids { get; set; }
+        public List<int> kids { get; set; }
         public string type { get; set; }
         public string url { get; set; }
         public int score { get; set; }
