+++ HackerNewsScraper/Models/CommentText.cs
@@ -5,7 +5,7 @@ namespace HackerNewsScraper.Models
 {
     public partial class CommentText
     {
-        public long CommentId { get; set; }
+        public int CommentId { get; set; }
         public string Text { get; set; }
 
         public virtual Comment Comment { get; set; }
