+++ HackerNewsScraper/Models/CommentUrl.cs
@@ -6,7 +6,7 @@ namespace HackerNewsScraper.Models
     public partial class CommentUrl
     {
         public int CommentUrlId { get; set; }
-        public long CommentId { get; set; }
+        public int CommentId { get; set; }
         public string Url { get; set; }
 
         public virtual Comment Comment { get; set; }
